+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubValidations (development version)
+
 # hubValidations 0.9.0
 
 * Re-exported functions useful for modelers (#149):