+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubValidations (development version)
+
 # hubValidations 0.10.0
 
 * Added `check_tbl_derived_task_id_vals()` check to `validate_model_data()`